sastify the test

+++ internal/integration_test/setup_test.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	_ "github.com/lib/pq"
 	"github.com/manabie-com/togo/config"
 	"github.com/manabie-com/togo/internal/services"
 	"github.com/manabie-com/togo/internal/storages"
@@ -21,10 +22,6 @@ func TestMain(m *testing.M) {
 }
 
 func createServices(m *testing.M) int {
-	// db, err := sql.Open("sqlite3", "../../data.db")
-	// if err != nil {
-	// 	log.Fatalf("error opening db %v", err)
-	// }
 
 	dbconnecter, err := config.GetDBConnecter()
 	if err != nil {
@@ -46,22 +43,60 @@ func createServices(m *testing.M) int {
 	return m.Run()
 }
 
-func truncate(db *sql.DB) error {
+func refreshUser(db *sql.DB, user *storages.User) error {
 	tx, err := db.Begin()
 	if err != nil {
 		return err
 	}
 
-	stmt := `DELETE FROM users`
-	if _, err := tx.Exec(stmt); err != nil {
+	stmt := `DELETE FROM tasks WHERE user_id=$1`
+	if _, err := tx.Exec(stmt, user.ID); err != nil {
 		tx.Rollback()
 		return err
 	}
 
-	stmt = `DELETE FROM tasks`
-	if _, err := tx.Exec(stmt); err != nil {
+	stmt = `DELETE FROM users WHERE id=$1`
+	if _, err := tx.Exec(stmt, user.ID); err != nil {
 		tx.Rollback()
 		return err
+	}
+
+	if err := tx.Commit(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func refreshSingleTask(db *sql.DB, task *storages.Task) error {
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+
+	stmt := `DELETE FROM tasks WHERE id=$1`
+	if _, err := tx.Exec(stmt, task.ID); err != nil {
+		tx.Rollback()
+		return err
+	}
+
+	if err := tx.Commit(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func refreshTasks(db *sql.DB, tasks []storages.Task) error {
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+
+	for _, task := range tasks {
+		stmt := `DELETE FROM tasks WHERE id=$1`
+		if _, err := tx.Exec(stmt, task.ID); err != nil {
+			tx.Rollback()
+			return err
+		}
 	}
 
 	if err := tx.Commit(); err != nil {
@@ -77,7 +112,7 @@ func seedUser(db *sql.DB) (*storages.User, error) {
 		Password: "example",
 	}
 
-	stmt := `INSERT INTO users (id, password) VALUES (?, ?)`
+	stmt := `INSERT INTO users (id, password) VALUES ($1, $2)`
 
 	_, err := db.Exec(stmt, &u.ID, &u.Password)
 	if err != nil {
@@ -94,11 +129,15 @@ func seedTaskItems(db *sql.DB, u *storages.User) ([]storages.Task, error) {
 			Content:     "write content here",
 			UserID:      u.ID,
 			CreatedDate: now.Format("2006-01-02"),
+			// CreatedDate: sql.NullTime{
+			// 	Time:  now,
+			// 	Valid: true,
+			// },
 		},
 	}
 
 	for _, task := range taskItems {
-		stmt := `INSERT INTO tasks (id, content, user_id, created_date) VALUES (?, ?, ?, ?)`
+		stmt := `INSERT INTO tasks (id, content, user_id, created_date) VALUES ($1, $2, $3, $4)`
 		_, err := db.Exec(stmt, &task.ID, &task.Content, &task.UserID, &task.CreatedDate)
 		if err != nil {
 			return nil, err
